[chore] Token 관련 로직을 Auth 폴더로 분리
