+++ src/main/java/com/wayble/server/auth/dto/TokenResponseDto.java
@@ -1,3 +1,3 @@
-package com.wayble.server.user.dto.token;
+package com.wayble.server.auth.dto;
 
 public record TokenResponseDto(String accessToken, String refreshToken) {}
