+++ src/main/java/com/wayble/server/auth/service/KakaoLoginService.java
@@ -1,4 +1,4 @@
-package com.wayble.server.user.service.auth;
+package com.wayble.server.auth.service;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
